Backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
httpx==0.19.0
beautifulsoup4==4.10.0
cryptography==36.0.0
motor==2.5.0
slowapi==0.1.5
playwright==1.22.0
httpx
pandas
scikit-learn
joblib
numpy
aiofiles
sqlalchemy